requirements.txt
paho-mqtt>=1.6.1
PyQt5>=5.15.2
PyQtWebEngine>=5.15.2
QT-PyQt-PySide-Custom-Widgets>=1.0.2
PySide6==6.9.1
python-dotenv>=0.21.0
PyQtChart>=5.15.2
# Dependências para previsão de séries temporais
pandas>=1.5.0
numpy>=1.21.0
statsmodels>=0.14.0